backend/requirements.txt
# Backend Dependencies (No Auth Version)

fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
aiofiles==24.1.0
pillow==11.0.0
redis==5.2.0
celery==5.4.0
boto3==1.35.59  # For S3 storage (optional)
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0